+++ earth2mip/models/diagnostic/base.py
