+++ earth2mip/models/diagnostic/precipitation_afno.py
