+++ earth2mip/models/diagnostic/wind_speed.py
