+++ test/diagnostics/test_climatenet.py
@@ -16,7 +16,7 @@
 import pytest
 import torch
 
-from earth2mip.models.dx import ClimateNet
+from earth2mip.models.diagnostic import ClimateNet
 
 
 @pytest.mark.slow
