+++ test/diagnostics/test_diagnostic_loop.py
@@ -19,8 +19,8 @@ import pytest
 import torch
 
 from earth2mip.grid import equiangular_lat_lon_grid
-from earth2mip.models.dx import DiagnosticTimeLoop, WindSpeed
-from earth2mip.models.dx.time_loop import filter_channels
+from earth2mip.models.diagnostic import DiagnosticTimeLoop, WindSpeed
+from earth2mip.models.diagnostic.time_loop import filter_channels
 from earth2mip.networks import Inference
