+++ test/diagnostics/test_precipitation.py
@@ -16,7 +16,7 @@
 import pytest
 import torch
 
-from earth2mip.models.dx import PrecipitationAFNO
+from earth2mip.models.diagnostic import PrecipitationAFNO
 
 
 @pytest.mark.slow
